requirements.txt
alembic==1.10.2
anyio==3.6.2
asgi-correlation-id==4.1.0
attrs==22.2.0
black==23.1.0
certifi==2022.12.7
click==8.1.3
coverage==7.2.1
dnspython==2.3.0
email-validator==1.3.1
exceptiongroup==1.1.0
fastapi==0.94.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
mypy-extensions==1.0.0
mysqlclient==2.1.1
orjson==3.8.7
packaging==23.0
pathspec==0.11.0
platformdirs==3.1.1
pluggy==1.0.0
pydantic==1.10.6
pyhumps==3.8.0
pytest==7.2.2
pytest-cov==4.0.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==2.0.5.post1
starlette==0.26.0.post1
tomli==2.0.1
typing_extensions==4.5.0
ujson==5.7.0
uvicorn==0.21.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
